Adds output file argument to fetcher daemon
When the fetcher is run through cron, the directory
it is run from is ~. Before, it was expected to be run
from ~/current, so the output (news.shelf) was not being
written to the right place. Now it takes a command line
argument for the output file. The crontab specifies
~/current/news.shelf, and everything works.