Finishes title bar; it now works.
The title bar now holds a link to the original article, or the default
text on the homepage (like, "Headliner - The Daily news"). For some
reason, the "content" and "text" divs seemed to steal all of the focus
from "pageTitle", so I put pageTitle *after* them on the page. This fixed
the focus stealing, but I'm not really sure why it was happening in the
first place. Still, cool, yeah? I'm really happy that it works.