Removed debugging prints, they weren't necessary
The problem i was having was that not every article in the database
had a timestamp value, which was weird. I realized that I had switched
over to the new, timestamp-indexed database before I updated the fetcher
code to create timestamps, so there were some old articles without them
that were still stuck in the database. I recreated the database, and now
everything works just fine because all the article objects have the same
'schema'.